Dedup the Integriforce rules.
Now that secadm uses a red-black tree that keys off of jail ID +
mountpoint + inode, hardlinking isn't supported. Hardlinking causes
hash collisions. Though this isn't really a problem per-se with PaX
settings, this is a problem with Integriforce.

Now, if a ruleset contains multiple rules referencing the same
underlying inode, use only the first rule. This means that the path
reported in the warning or error message cannot be guaranteed to be
correct. This allows a user to mass-hash all executables on a
filesystem, adding them all to Integriforce.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>